코드정리 및 선택삭제(차량,게시판)

+++ src/main/java/com/kh/rent/admin/controller/AdBoardController.java
@@ -1,15 +1,13 @@
 package com.kh.rent.admin.controller;
 
 
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.kh.rent.admin.service.AdBoardService;
@@ -18,7 +16,7 @@ import lombok.extern.log4j.Log4j;
 
 @Controller
 @Log4j
-@RequestMapping("/admin/board/")
+@RequestMapping("/admin/board/*")
 public class AdBoardController {
 	
 	@Autowired
@@ -26,16 +24,27 @@ public class AdBoardController {
 
     @PostMapping("/deleteCheckedBoard")
     @ResponseBody
-    public String deleteCheckedBoard(@RequestParam List<Integer> boardNumbers) {
+    public String deleteCheckedBoard(@RequestBody Map<String, Object> requestBody) {
         try {
-            int count = boardService.deleteCheckedBoard(boardNumbers);
-            if (count == 1) {
-                return "Success";
+            @SuppressWarnings("unchecked")
+			List<Integer> boardNumbers = (List<Integer>) requestBody.get("boardNumbers");
+            if (boardNumbers != null) {
+                log.info("boardNumbers: " + boardNumbers);
+
+                // boardNumbers 파라미터가 올바르게 전달되었는지 로깅합니다.
+                int count = boardService.deleteCheckedBoard(boardNumbers);
+                log.info("count"+count);
+                if (count > 0) {
+                    return "success";
+                } else {
+                    return "fail";
+                }
+                
             } else {
-                return "fail";
+                return "Error";
             }
         } catch (Exception e) {
-            e.printStackTrace();
+            log.error("Error occurred:", e);
             return "Error";
         }
     }
